+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 2.0.0
 
 * Extend list functionality:
-  * methods: `remove`
+  * methods: `asMapOfLists`, `remove`
   * Rename `remove` to `removeFirst`
 * Extend set functionality:
   * methods: `addList`